Add replication and database schema
Add docker-compose.yaml for easy management of multiple containers.
It's currently setup for physical replication with one primary database and one
replication database. It's operating on an inidividual network with data
persistence, possible by using docker volumes.

As the name suggests init-scipts/ directory holds the scripts for
initializing the replication process as well as creating the schema of
the whole database and filling it with predefined entries.

Overview and order:

1. docker-compose.yaml - starts the containers
2. setup-replication.sh - readies the primary database for replication
3. setup-database.sql - database schema and filling
4. replica-entrypoint.sh - checks if a database is already created. If
   not - it creates the replication database and establishes a
   connection with primary database.